Refactor Formatter to UI as it makes more sense to be under UI. Refactor Welcome screen to Formatter
