+++ src/main/java/seedu/duke/command/AddEventCommand.java
@@ -3,7 +3,7 @@ package seedu.duke.command;
 import seedu.duke.data.timetable.Event;
 import seedu.duke.data.timetable.RecurringEvent;
 import seedu.duke.util.DateTimeManager;
-import seedu.duke.util.Formatter;
+import seedu.duke.ui.Formatter;
 
 import static seedu.duke.util.PrefixSyntax.PREFIX_DELIMITER;
 import static seedu.duke.util.PrefixSyntax.PREFIX_RECURRING;
