+++ src/main/java/seedu/duke/command/CreateTagCommand.java
@@ -2,7 +2,7 @@ package seedu.duke.command;
 
 import seedu.duke.data.notebook.Tag;
 import seedu.duke.data.notebook.Tag.TagColor;
-import seedu.duke.util.Formatter;
+import seedu.duke.ui.Formatter;
 
 import java.util.ArrayList;
