+++ src/main/java/seedu/duke/command/HelpCommand.java
@@ -1,7 +1,7 @@
 package seedu.duke.command;
 
 import seedu.duke.data.notebook.Tag.TagColor;
-import seedu.duke.util.Formatter;
+import seedu.duke.ui.Formatter;
 
 import java.io.IOException;
 import java.util.logging.ConsoleHandler;
