+++ src/main/java/seedu/duke/command/ListNoteCommand.java
@@ -2,7 +2,7 @@ package seedu.duke.command;
 
 import seedu.duke.data.notebook.Note;
 import seedu.duke.data.notebook.Tag;
-import seedu.duke.util.Formatter;
+import seedu.duke.ui.Formatter;
 
 import java.util.Collections;
 import java.util.Comparator;
