+++ src/main/java/seedu/duke/data/notebook/TagManager.java
@@ -1,6 +1,6 @@
 package seedu.duke.data.notebook;
 
-import seedu.duke.util.Formatter;
+import seedu.duke.ui.Formatter;
 
 import java.util.ArrayList;
 import java.util.HashMap;
