+++ src/main/java/seedu/duke/ui/Formatter.java
@@ -1,4 +1,4 @@
-package seedu.duke.util;
+package seedu.duke.ui;
 
 import seedu.duke.data.notebook.Notebook;
 import seedu.duke.data.notebook.Note;
@@ -15,6 +15,33 @@ public class Formatter {
     private static final String ROW_SPLIT = "-";
     private static final String COLUMN_SPLIT = "|";
     private static final int MAX_CONTENT_LENGTH = 50;
+
+    // Character code adapted from http://patorjk.com/software/taag/#p=display&f=Ghost&t=NotUS
+    // Slight modifications made to make it easier on the eyes
+    private static final String NOTUS_LOGO = Formatter.LS
+            + Formatter.LS
+            + "     .-') _               .-') _                 .-')    "
+            + Formatter.LS
+            + "    ( OO ) )             (  OO) )               ( OO ).  "
+            + Formatter.LS
+            + ",--./ ,--,'  .-'),-----. /     '._ ,--. ,--.   (_)---\\_) "
+            + Formatter.LS
+            + "|   \\ |  |\\ ( OO'  .-.  '|'--...__)|  | |  |   /    _ |  "
+            + Formatter.LS
+            + "|    \\|  | )/   |  | |  |'--.  .--'|  | | .-') \\  :` `.  "
+            + Formatter.LS
+            + "|  .     |/ \\_) |  | |  |   |  |   |  | |( OO ) '..`''.) "
+            + Formatter.LS
+            + "|  |\\    |    \\ |  | |  |   |  |   |  | | `-' /.-._)   \\ "
+            + Formatter.LS
+            + "|  | \\   |     `'  '-'  '   |  |  ('  '-'(_.-' \\       / "
+            + Formatter.LS
+            + "`--'  `--'       `-----'    `--'    `-----'     `-----'  "
+            + Formatter.LS;
+
+    public static String getNotusLogo() {
+        return NOTUS_LOGO;
+    }
 
     public static String formatNotebook(Notebook notebook) {
         String formattedString = "";
