+++ src/test/java/seedu/duke/command/DeleteTagCommandTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import seedu.duke.data.notebook.Note;
 import seedu.duke.data.notebook.TagManager;
 import seedu.duke.data.notebook.Tag;
-import seedu.duke.util.Formatter;
+import seedu.duke.ui.Formatter;
 
 import static seedu.duke.command.DeleteTagCommand.COMMAND_SUCCESSFUL_MESSAGE;
 import static seedu.duke.command.DeleteTagCommand.COMMAND_UNSUCCESSFUL_MESSAGE;
